Solved the problem of printed 'return'
I made a change on the way we read and print input given by the user, to
avoid the need of the extra bottom line. I stripped the program of the
built-in autocomplete functionality when opening a new torrent. so now I
have to write a custom one in the following commits.